no longer need separate v4.compose
put it back in nbbase, and work around circular import another way
