+++ IPython/nbformat/v4/tests/nbexamples.py
@@ -3,7 +3,7 @@
 import os
 from base64 import encodestring
 
-from ..compose import (
+from ..nbbase import (
     new_code_cell, new_heading_cell, new_markdown_cell, new_notebook,
     new_output, new_raw_cell
 )
